Centralise modal test suite
