+++ lib/components/Dialog/Dialog.test.ts
@@ -1,0 +1,4 @@
+import { modalTestSuite } from '../private/Modal/modalTestSuite';
+import { Dialog } from '..';
+
+modalTestSuite('Dialog', Dialog);
