+++ lib/components/Drawer/Drawer.test.ts
@@ -1,0 +1,4 @@
+import { modalTestSuite } from '../private/Modal/modalTestSuite';
+import { Drawer } from '..';
+
+modalTestSuite('Drawer', Drawer);
